modules/recon.py
from typing import Dict, Any, List
import dns.resolver
import requests
from concurrent.futures import ThreadPoolExecutor
from .base import BaseModule

class ReconModule(BaseModule):
    def __init__(self):
        super().__init__()
        self.subdomains: List[str] = []
        self.assets: Dict[str, Any] = {}
        
    def run(self, target: str, options: Dict[str, Any]) -> Dict[str, Any]:
        self.logger.info(f"Starting reconnaissance on {target}")
        
        if options.get('scan_subdomains', True):
            self.scan_subdomains(target)
        
        if options.get('discover_assets', True):
            self.discover_assets(target)
            
        return {
            'subdomains': self.subdomains,
            'assets': self.assets
        }
    
    def scan_subdomains(self, target: str):
        """Perform subdomain enumeration"""
        wordlist = self._load_wordlist('subdomains.txt')
        with ThreadPoolExecutor(max_workers=10) as executor:
            futures = [executor.submit(self._check_subdomain, f"{word}.{target}") 
                      for word in wordlist]
            
            for future in futures:
                if subdomain := future.result():
                    self.subdomains.append(subdomain)
    
    def discover_assets(self, target: str):
        """Discover web assets"""
        self._scan_common_files(target)
        self._scan_technologies(target)
        self._scan_certificates(target)
    
    def _check_subdomain(self, subdomain: str) -> str:
        try:
            dns.resolver.resolve(subdomain, 'A')
            return subdomain
        except:
            return ""
    
    def _load_wordlist(self, filename: str) -> List[str]:
        # TODO: Implement wordlist loading
        return ['www', 'dev', 'staging', 'test']
    
    def _scan_common_files(self, target: str):
        common_files = ['robots.txt', 'sitemap.xml', '.git']
        for file in common_files:
            try:
                resp = requests.get(f"https://{target}/{file}", timeout=5)
                if resp.status_code == 200:
                    self.assets[file] = resp.url
            except:
                continue
    
    def _scan_technologies(self, target: str):
        # TODO: Implement technology detection (e.g., Wappalyzer-like)
        pass
    
    def _scan_certificates(self, target: str):
        # TODO: Implement SSL/TLS certificate scanning
        pass
