requirements.txt
colorama==0.4.6
dnspython>=2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
reportlab==4.0.4
tabulate==0.9.0
tqdm==4.65.0
ipaddress==1.0.23
markdown==3.4.3
scapy==2.5.0
aiohttp==3.8.5
cryptography==41.0.3
python-nmap==0.7.1
PyJWT==2.8.0
SQLAlchemy==2.0.20
python-whois==0.8.0
boto3==1.28.36
azure-mgmt-compute==30.0.0
google-cloud-compute==1.13.0
